+++ back/Giboo/src/main/java/kh/fin/giboo/member/model/service/MyPageService.java
@@ -1,0 +1,8 @@
+package kh.fin.giboo.member.model.service;
+
+import java.util.Map;
+
+public interface MyPageService {
+
+	int updateInfo(Map<String, Object> paramMap);
+}
