+++ back/Giboo/src/main/java/kh/fin/giboo/member/model/service/MyPageServiceImpl.java
@@ -1,0 +1,26 @@
+package kh.fin.giboo.member.model.service;
+
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+
+import kh.fin.giboo.member.model.dao.MyPageDAO;
+
+@Service
+public class MyPageServiceImpl implements MyPageService {
+
+	@Autowired
+	private MyPageDAO dao;
+	
+	@Autowired
+	private BCryptPasswordEncoder bcrypt;
+	
+	//회원정보 서비스 구현
+	@Override
+	public int updateInfo(Map<String, Object> paramMap) {
+		
+		return dao.updateInfo(paramMap);
+	}
+}
