+++ back/Giboo/src/main/java/kh/fin/giboo/member/model/vo/Myactive.java
@@ -1,0 +1,21 @@
+package kh.fin.giboo.member.model.vo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+
+
+public class Myactive {
+
+	//나의 활동
+	private int myactNo;
+	private String volcomStatus;
+	private int donationNo;
+	private int volunteerNo;
+	private int memberNo;
+	private int eventNo;
+}
