--- back/Giboo/src/main/java/kh/fin/giboo/mypage/model/dao/MyPageDAO.java
@@ -1,21 +1,0 @@
-package kh.fin.giboo.mypage.model.dao;
-
-import java.util.Map;
-
-import org.mybatis.spring.SqlSessionTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MyPageDAO {
-	@Autowired
-	private SqlSessionTemplate sqlSession;
-	
-	/**회원정보 수정 DAO
-	 * @param paramMap
-	 * @return result
-	 */
-	public int updateInfo(Map<String, Object> paramMap) {	
-		return sqlSession.update("myPageMapper.updateInfo", paramMap);
-	}
-}
