--- back/Giboo/src/main/java/kh/fin/giboo/mypage/model/service/MyPageService.java
@@ -1,8 +1,0 @@
-package kh.fin.giboo.mypage.model.service;
-
-import java.util.Map;
-
-public interface MyPageService {
-
-	int updateInfo(Map<String, Object> paramMap);
-}
