--- back/Giboo/src/main/java/kh/fin/giboo/mypage/model/service/MyPageServiceImpl.java
@@ -1,26 +1,0 @@
-package kh.fin.giboo.mypage.model.service;
-
-import java.util.Map;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import kh.fin.giboo.mypage.model.dao.MyPageDAO;
-
-@Service
-public class MyPageServiceImpl implements MyPageService {
-
-	@Autowired
-	private MyPageDAO dao;
-	
-	@Autowired
-	private BCryptPasswordEncoder bcrypt;
-	
-	//회원정보 서비스 구현
-	@Override
-	public int updateInfo(Map<String, Object> paramMap) {
-		
-		return dao.updateInfo(paramMap);
-	}
-}
