--- back/Giboo/src/main/java/kh/fin/giboo/mypage/model/vo/myactive.java
@@ -1,21 +1,0 @@
-package kh.fin.giboo.mypage.model.vo;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-
-
-public class myactive {
-
-	//나의 활동
-	private int myactNo;
-	private String volcomStatus;
-	private int donationNo;
-	private int volunteerNo;
-	private int memberNo;
-	private int eventNo;
-}
